lucet-runtime: DlModule load now optionally enforces version match
this check exists to make sure that lucetc and lucet-runtime are built
against the same revision of lucet-module - the use of serde depends
on both sides defining types the same.

this check can be overridden for cases where the user's runtime needs to
load modules built with a different lucetc. The user is responsible for
ensuring that the modules were created with the same lucet-module type
definitions.