use wasi sdk 12 (#615)
* lucet-wasi-sdk: wasi sdk 12 no longer supports `--no-threads`

* dockerfile: use wasi-sdk 12

* ci: use wasi-sdk 12

* lucet-runtime: fix tests for wasi-sdk-12 by rewriting use_allocator.c in wat

use_allocator.c was a fixture used for tests that manipulated the size
of linear memory. written in C, it depended heavily on the implemention
of wasi-libc's allocator, which changed from sdk-11 to sdk-12. Rather
than try to depend on the new (still fragile) behavior these tests were
rewritten in pure wat, with a much simpler allocator behavior that just
expands the linear memory. The tests were modified to observe this new
behavior.