Add new x64 backend as lucetc option, and add to CI runs.
The new x86-64 backend in Cranelift is now able to run all Lucet tests,
and has shown better performance (compile-time and run-time) than the
current backend. This PR adds a Cargo feature to lucetc to build with
the new backend, and modifies the Makefile and CircleCI config to run
tests with the new backend as well, in order to ensure that it continues
to work.