Fix some stack overflow-related issues exposed by the new-backend tests.
This adjusts some tests that had specific asserts on stack-overflow
limits, based on differences in the new backend's stack frame allocation
(specifically, every local gets an 8-byte-aligned slot, even if it is an
i32). The adjusted tests should work on both old and new backends.

This also pulls in a new wasmtime that has some fixes for
probestack-related issues (bytecodealliance/wasmtime#2486).