Allow yielding from an async execution context
By replacing the expected yield value from a Box<PhantomData<T>> to
TypeId, it becomes possible to resume instance execution when the size
of a boxed type is unknown and cannot be passed as an argument.

This makes it possible to yield and resume from an async context by
passing the resumption value through RunAsyncState.

Finally, we have to check that the context didn't change between
resumption within try_block_on, in order to make sure the right Waker
is scheduled and the future will be resumed when it is ready.