lucetc: bump reserved-size default value
Set the reserved-size default value to 4 GiB, so that it aligns
with the runtime's defaults, and provides good performance out of
the box.

Also the individual `--min-reserved-size` and `--max-reserved-size`
knobs from the CLI to be less intimidating.

They remain accessible programatically.

Update the documentation by the way to include all the command-line
flags that are currently supported.

Fixes #607