Make the new Cranelift x64 backend the default.
After sufficient testing, we're confident that the new x64 backend does
not have any correctness issues relative to the old backend; and because
it generates better code, we would like to switch the default. This PR
will cause `lucetc` to use the new backend by default. Building `lucetc`
with the `old-x64-backend` Cargo feature will swap the old backend in
again. This PR also modifies the new-backend CI tests to test the old
backend instead (to ensure that it keeps working as an alternative,
at least for a while).