vendor wiggle and wasi-common crates from alexcrichton/new-api
The wasmtime new-api branch constains futures to be Send, and wiggle and
wasi-common got considerable changes to fix this. Lucet futures are Send
as well, so using these versions of wiggle and wasi-common solves a
bunch of integration issues.

We expect this branch to take a little bit of time to land because it is
gated behind an RFC, so vendoring these changes allows us to ship the async
wasi-common in Lucet a lot sooner.