Integrate VeriWasm.
This commit brings in VeriWasm, a static analysis tool that
independently verifies the sound heap-sandboxing and other properties of
machine code generated by Cranelift as used by Lucet. It adds an option
`--veriwasm` to the compiler `lucetc` so that, in concert with changes
recently made to VeriWasm and Cranelift, we can verify generated code in
memory, right after we generate it, and with reasonable overhead
(measured at ~30% additional compile time for some simple test cases).